refactor artpiece preview as a component
