+++ styles.js
@@ -30,4 +30,7 @@ a {
   text-decoration: none;
   color: inherit;
 }
+img {
+    pointer-events: none;
+}
 `;
